requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.5.0
Pillow==9.5.0
plotly==5.14.0
streamlit==1.20.0
scikit-learn==1.0.2
